emotion-music-player/services/api/pyproject.toml
[project]
name = "emotion-music-player-api"
version = "0.1.0"
description = "FastAPI backend for emotion music player"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg2-binary>=2.9.9",
  "requests>=2.32.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.2.0",
  "pydantic-settings>=2.3.4"
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true